requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
googlesearch-python==1.2.5
idna==3.8
requests==2.32.3
soupsieve==2.6
urllib3==2.2.2
